flask-server/requirements.txt
Flask==2.2.3
werkzeug
llama_index
pathlib
langchain
pandas
numpy
wheel
python-dotenv
gpt-index
PyPDF2
PyCryptodome
unstructured
chromadb
python-magic-bin
nltk
magic
openai
